requirements.txt
python-telegram-bot==20.7
rapidfuzz==3.5.2
openai==1.3.8
aiohttp==3.9.1
